Create a function called playerChoice
i moved my variable into this function, so
instead of the prompt popping up on the
page refresh. it pops up when the function is
called.

this is helpful because i can place this new
function inside my game function, that way it
only ask's for input when playing the game.